rksfunPalettes for plotting with ggplot2

Function for using rksfunPalettes colors schemes in ggplot2. Use scale_color_rksfun_d and scale_fill_rksfun_d for discrete scales and scale_color_rksfun_c and scale_fill_rksfun_c for continuous scales.

Usage

scale_color_rksfun_d(name, direction = 1, override.order = FALSE, ...)

Arguments

name

Name of Palette. Choices are: rktbns_post, and tf2

direction

Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed

override.order

Colors are picked from palette to maximize readability and aesthetics. This means that colors are not always selected in sequential order from the full palette. If override.order is set to TRUE, colors are selected in sequential order from the full palette instead. Default is FALSE.

...

Other arguments passed on to discrete_scale

Value

A function that returns a discrete color scale.